requirements.txt
Flask
pytesseract
pdf2image
Pillow
opencv-python
numpy
spacy
scikit-learn
joblib
Werkzeug
gunicorn
requests
flask-cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0